fix(chain)!: fix `KeychainTxOutIndex` range methods
Previously, keychain (`K`) range queries did to translate well to
`DescriptorId` queries (which is used in the internal `SpkTxOutIndex`).
This is because the `Ord` of `K` is not guaranteed to be one-to-one with
the `Ord` of the corresponding `DescriptorId`.

The solution (as introduced here), is to range over the
`keychains_to_descriptor_ids` field and find corresponding
`DescriptorId`s first. Then query the internal index with the
`DescriptorId`s individually.