feat(chain)!: Change canonicalization behavior for unanchored and unseen txs
Previously, transactions without anchors or last_seen values can only be
considered canonical by means of transitivity (one of their descendants
are deemed canonical).

This is changed so that txs without anchors or last_seen can be
canonical as long as they do not conflict with any other transaction
already determined canonical.

Why? This is the better default behavior as we want to avoid
accidentally replacing transactions. Transaction replacements should be
explicit.

TODO: There is no sane way to prioritize between two txs that are both
unseen and unanchored. Consider introducing an `added_seq` which is a
sequence number that increments everytime a tx is added to the tx graph.
The sequence will be persisted and forever associated with that tx.

BREAKING: Signature of `CanoncialReason` and it's `descendant` method
are changed.