feat(chain)!: Change `TxGraph` to understand missing-at & last-missing
The missing-at and last-missing timestamp informs the `TxGraph` when the
transaction was last deemed as missing from the mempool (evicted).

The canonicalization algorithm is changed to disregard transactions with
a last-missing timestamp greater or equal to it's last-seen timestamp.
The exception is when we have a canonical descendant due to rules of
transitivity.